Restructure files and make all world subclasses library
