+++ src/filters/BallFilter.cpp
@@ -2,8 +2,8 @@
 // Created by rolf on 17-11-19.
 //
 
-#include "BallFilter.h"
-#include "util/Scaling.h"
+#include "filters/BallFilter.h"
+#include "Scaling.h"
 
 BallFilter::BallFilter(const proto::SSL_DetectionBall &detectionBall, double detectTime, int cameraID) :
         CameraFilter(detectTime, cameraID),
