+++ src/filters/Scaling.h
